Change "find()" of AzureBlobFileSystem to match gcsfs implementation
Previously the find would take a path and recursively list all the
sub folders until maxdepth(default to None) is reached, this was mostly
used by glob().

This changes the behavior to the following one:

Take a path (a directory) and a prefix (partial match)
and list ervery blob that matches "^{path}/{prefix}*"
This is done in one single API call, no recursions.
The `maxdepth` argument of "find" has been dropped since it doesn't have
any meaning now all the matching blobs will be returned.

It suffers the same 'bug' regarding withdirs kwargs:

  The following structure might bring wrong dir results:
    - directory_a/a/b/c/file.txt
  when querying path: `directory_a` prefix: `a` and withsubdirs
  only directory_a/b/c will be returned as a directory which means
  that only folder with a file will be returned, Reading the from
  gcsfs, it looks like this is normal behavior.
  It might not be considered a but thou

The glob function behavior however remains unchanged since this is the
only way we can do proper complex pattern matching with branch elimination
which might, for some, be useful.