Haneld the situation that Azure Datalake Gen1 returns filename contains special sub-string
AzureDatalakeFileSystem.ls has special mechanism for type and length.
But the above special mechanism should only applicable when call ls
with detail=True and ls returns a list of dictionary.

When call ls with detail=False, it returns a list of str. Keyword in
for str has the meaning of check if one str is substring of another.
That is not what the mechanism should acting on.