Respect onerror='return' in AzureBlobFileSystem.cat
This commits adds an option to keep paths in the output set of
AzureBlobFileSystem._expand_path even when the blob doesn't exist.

This change is motivated by the onerror='return' option in cat, one of
the callers of _expand_path (or expand_path). The stripping of
none-existing paths undermines the onerror='return' and onerror='omit'
in cat.