Introduce connect client tests
This builds off of the tests that asserted adlfs user agents are set
for SDK clients. Even though the user agent tests added some coverage
to how adlfs connected to SDK coverage, the adlfs test suite lacked
coverage for how adlfs configurations were proxied to SDK clients
(e.g., which account url to use and what credentials to set). By
having these tests, it will make it easier to correctly refactorings
related to client creation and easier to add support new SDK client
configurations in the future.

In addition to add these tests, a bug was discovered where for SAS
token authentication on the file-like object, it would not respect
location mode, while the file system did. The logic in the file-like
object was updated to be consistent with files system logic.