Respect protocol tuple when stripping full URIs
This change makes AzureBlobFileSystem consistent with other fsspec
implementations that use the protocol class tuple when stripping
the protocol from fully qualified URIs. It also unblocks consumers
to be able to update this protocol tuple if out of the box adlfs
does not respect a particular protocol (e.g., deprecated wasb://).